New cypress support utils function
